delete 기능 get 삭제

+++ app/src/routes/arte/arte.ctrl.js
@@ -87,22 +87,6 @@ const output = {
             }
         });    
     },
-    
-    getDelete: (req, res) => {
-        const boardno = parseInt(req.params.boardno);
-        const nickname = req.session.nickname;
-        const query = "SELECT * FROM arteboard where BOARD_NO = ?";
-        db.query(query, [boardno], (err, result) => {
-            if (err) console.log(err);
-            if (result) {
-                if (result[0].nickname !== nickname) {
-                    res.send("<script>alert('본인이 작성한 글만 삭제할 수 있습니다.');location.href=history.back();</script>");
-                } else {
-                    res.render("arte/arteDelete", {'nickname':nickname});
-                }
-            }
-        }); 
-    },  
     getApply: (req, res) => {
         res.send("<script>alert('신청이 완료되었습니다.');location.href='/arte';</script>");
     },
@@ -118,9 +102,6 @@ const output = {
             });
         }   
     },
-    getmusic: (req, res) => {
-        
-    }
 }
 
 const process = {
@@ -159,11 +140,21 @@ const process = {
     postDelete: (req, res) => {
         const boardno = parseInt(req.params.boardno);
         const nickname = req.session.nickname;
-        const query = "DELETE FROM arteboard WHERE BOARD_NO = ?;";
-        db.query(query, [boardno], (err, result) => {
-            if (err) return console.log(err);
-            if (result) res.json({success: true});
-        });
+        const query1 = "SELECT * FROM arteboard where BOARD_NO = ?";
+        db.query(query1, [boardno], (err, result) => {
+            if (err) console.log(err);
+            if (result) {
+                if (result[0].nickname !== nickname) {
+                    res.send("<script>alert('본인이 작성한 글만 삭제할 수 있습니다.');location.href=history.back();</script>");
+                } else {
+                    const query2 = "DELETE FROM arteboard WHERE BOARD_NO = ?;";
+                    db.query(query2, [boardno], (err, result) => {
+                        if (err) return console.log(err);
+                        if (result) res.json({success: true});
+                    });
+                }
+            }
+        }); 
     },
     postApply: (req, res) => {
         const boardno = parseInt(req.params.boardno);
